agregar y eliminar carreras y esceulas

+++ src/main/java/com/proyecto/SsYPp/Service/Impl/EscuelasImpl.java
@@ -5,6 +5,7 @@ import com.proyecto.SsYPp.Repository.EscuelasRepository;
 import com.proyecto.SsYPp.Service.EscuelasService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import com.proyecto.SsYPp.Repository.CarreraRepository;
 
 import java.util.List;
 
@@ -13,6 +14,9 @@ public class EscuelasImpl implements EscuelasService {
 
     @Autowired
     private EscuelasRepository escuelasRepository;
+
+    @Autowired
+    private CarreraRepository carreraRepository;
 
     @Override
     public List<Escuelas> obtenerEscuelas() {
@@ -27,6 +31,16 @@ public class EscuelasImpl implements EscuelasService {
     }
     @Override
     public void eliminarEscuela(Integer id) {
+
+        boolean tieneCarreras =
+                carreraRepository.existsByEscuela_Idescuela(id);
+
+        if (tieneCarreras) {
+
+            throw new RuntimeException(
+                    "No se puede eliminar la escuela porque tiene carreras registradas."
+            );
+        }
 
         escuelasRepository.deleteById(id);
     }
